fix bug #45

+++ app/src/main/java/com/ostrovec/mygarden/utils/CalendarWorker.kt
@@ -3,7 +3,7 @@ package com.ostrovec.mygarden.utils
 class CalendarWorker {
     companion object {
         fun convertDaysToMilliseconds(days: Int): Long {
-            return (days * 1000 * 60 * 60 * 24).toLong()
+            return days * 1000L * 60 * 60 * 24
         }
 
         fun convertMillisecondsInDays(milliseconds: Long): Int {
